+++ components/peripherals/peripherals/flow_meter/FlowMeterComponent.hpp
@@ -77,7 +77,7 @@ private:
         auto duration = duration_cast<microseconds>(lastMeasurement - lastPublished);
         if (duration > microseconds::zero()) {
             // Flow rate is measured in in liters / min
-            json["flowRate"] = currentVolume / duration.count() * 1000 * 1000 * 60;
+            json["flowRate"] = currentVolume / static_cast<double>(duration.count()) * 1000 * 1000 * 60;
         }
         lastPublished = lastMeasurement;
     }
